// snl-gms-mock/packages/common-model/src/ts/event/types.ts
export enum RestraintType {
  UNRESTRAINED = 'UNRESTRAINED',
  FIXED = 'FIXED'
}
// ...
export enum RestrainerType {
  FIXED_BY_ANALYST = 'FIXED_BY_ANALYST',
  FIXED_BY_CONFIGURATION = 'FIXED_BY_CONFIGURATION',
  FIXED_BY_LOCATOR = 'FIXED_BY_LOCATOR',
  UNKNOWN = 'UNKNOWN'
}

/**
 * Enumerated Depth Restraint Reason
 */
export enum DepthRestraintReason {
  FIXED_AT_DEPTH_FOUND_USING_DEPTH_PHASE_MEASUREMENTS = 'FIXED_AT_DEPTH_FOUND_USING_DEPTH_PHASE_MEASUREMENTS',
  FIXED_AT_STANDARD_DEPTH = 'FIXED_AT_STANDARD_DEPTH',
  FIXED_AT_SURFACE = 'FIXED_AT_SURFACE',
  OTHER = 'OTHER'
}